+++ spring-data-elasticsearch-rest/src/test/java/com/github/ydespreaux/spring/data/elasticsearch/ITSuiteTest.java
@@ -11,15 +11,9 @@ import org.junit.runners.Suite;
         ITReactiveElasticsearchTemplateTest.class,
         ITSampleEntityRepositoryTest.class,
         ITSampleEntityWithAliasRepositoryTest.class,
-        ITElasticsearchTemplateRolloverTest.class
+        ITElasticsearchTemplateRolloverTest.class,
+        ITElasticsearchTemplateParentFieldChildTest.class
 })
 public class ITSuiteTest {
 
-//    @ClassRule
-//    public static final ElasticsearchContainer elasticContainer = new ElasticsearchContainer("6.4.2")
-//            .withConfigDirectory("elastic-config");
-
-//    static {
-//        System.setProperty("spring.elasticsearch.rest.uris", "http://localhost:9200");
-//    }
 }